remove time-sensitive warning on executed action
